Updating invalidations to use new query keys
