+++ web-local/src/lib/components/leaderboard/LeaderboardMeasureSelector.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
   import { EntityStatus } from "@rilldata/web-local/common/data-modeler-state-service/entity-state-service/EntityStateService";
+  import { useMetaQuery } from "@rilldata/web-local/lib/svelte-query/dashboards";
   import { crossfade, fly } from "svelte/transition";
   import {
     MetricsExplorerEntity,
     metricsExplorerStore,
   } from "../../application-state-stores/explorer-stores";
-  import { useMetaQuery } from "../../svelte-query/queries/metrics-views/metadata";
   import { SelectMenu } from "../menu";
   import Spinner from "../Spinner.svelte";
   import type { MetricsViewMeasure } from "@rilldata/web-common/runtime-client";
