+++ web-local/src/lib/components/navigation/dashboards/MetricsDefinitionAssets.svelte
@@ -2,7 +2,6 @@
   import { goto } from "$app/navigation";
   import { page } from "$app/stores";
   import {
-    getRuntimeServiceListFilesQueryKey,
     useRuntimeServiceDeleteFileAndReconcile,
     useRuntimeServicePutFileAndReconcile,
   } from "@rilldata/web-common/runtime-client";
@@ -27,6 +26,7 @@
   import { Divider } from "@rilldata/web-local/lib/components/menu/index.js";
   import { deleteFileArtifact } from "@rilldata/web-local/lib/svelte-query/actions";
   import { useDashboardNames } from "@rilldata/web-local/lib/svelte-query/dashboards";
+  import { invalidateAfterReconcile } from "@rilldata/web-local/lib/svelte-query/invalidation";
   import { getContext } from "svelte";
   import { slide } from "svelte/transition";
   import type { ApplicationStore } from "../../../application-state-stores/application-store";
@@ -67,7 +67,7 @@
       showMetricsDefs = true;
     }
     const newDashboardName = getName("dashboard", $dashboardNames.data);
-    await $createDashboard.mutateAsync({
+    const resp = await $createDashboard.mutateAsync({
       data: {
         instanceId,
         path: `dashboards/${newDashboardName}.yaml`,
@@ -78,9 +78,7 @@
       },
     });
     goto(`/dashboard/${newDashboardName}`);
-    queryClient.invalidateQueries(
-      getRuntimeServiceListFilesQueryKey(instanceId)
-    );
+    return invalidateAfterReconcile(queryClient, instanceId, resp);
   };
 
   const editModel = (sourceModelName: string) => {
