+++ web-local/src/lib/components/navigation/models/createModel.ts
@@ -1,12 +1,10 @@
 import { goto } from "$app/navigation";
-import {
-  getRuntimeServiceListFilesQueryKey,
-  V1PutFileAndReconcileResponse,
-} from "@rilldata/web-common/runtime-client";
+import type { V1PutFileAndReconcileResponse } from "@rilldata/web-common/runtime-client";
 import { EntityType } from "@rilldata/web-local/common/data-modeler-state-service/entity-state-service/EntityStateService";
 import { getName } from "@rilldata/web-local/common/utils/incrementName";
 import { dataModelerService } from "@rilldata/web-local/lib/application-state-stores/application-store";
 import { fileArtifactsStore } from "@rilldata/web-local/lib/application-state-stores/file-artifacts-store";
+import { invalidateAfterReconcile } from "@rilldata/web-local/lib/svelte-query/invalidation";
 import { getFileFromName } from "@rilldata/web-local/lib/util/entity-mappers";
 import { queryClient } from "@rilldata/web-local/lib/svelte-query/globalQueryClient";
 import type { UseMutationResult } from "@sveltestack/svelte-query";
@@ -29,6 +27,7 @@ export async function createModel(
     },
   });
   fileArtifactsStore.setErrors(resp.affectedPaths, resp.errors);
+  invalidateAfterReconcile(queryClient, instanceId, resp);
   if (resp.errors?.length && sql !== "") {
     resp.errors.forEach((error) => {
       console.error(error);
@@ -40,9 +39,6 @@ export async function createModel(
   ]);
   if (!setAsActive) return;
   goto(`/model/${newModelName}`);
-  return queryClient.invalidateQueries(
-    getRuntimeServiceListFilesQueryKey(instanceId)
-  );
 }
 
 export async function createModelFromSource(
