+++ web-local/src/lib/components/navigation/sources/createSource.ts
@@ -1,12 +1,12 @@
 import { goto } from "$app/navigation";
-import {
-  getRuntimeServiceListFilesQueryKey,
+import type {
   V1PutFileAndReconcileResponse,
   V1ReconcileError,
 } from "@rilldata/web-common/runtime-client";
 import { EntityType } from "@rilldata/web-local/common/data-modeler-state-service/entity-state-service/EntityStateService";
 import { dataModelerService } from "@rilldata/web-local/lib/application-state-stores/application-store";
 import { fileArtifactsStore } from "@rilldata/web-local/lib/application-state-stores/file-artifacts-store";
+import { invalidateAfterReconcile } from "@rilldata/web-local/lib/svelte-query/invalidation";
 import { getFileFromName } from "@rilldata/web-local/lib/util/entity-mappers";
 import { queryClient } from "@rilldata/web-local/lib/svelte-query/globalQueryClient";
 import type { UseMutationResult } from "@sveltestack/svelte-query";
@@ -34,8 +34,6 @@ export async function createSource(
   }
   await dataModelerService.dispatch("addOrSyncTableFromDB", [tableName, true]);
   goto(`/source/${tableName}`);
-  await queryClient.invalidateQueries(
-    getRuntimeServiceListFilesQueryKey(instanceId)
-  );
+  invalidateAfterReconcile(queryClient, instanceId, resp);
   return [];
 }
