+++ web-local/src/lib/svelte-query/actions.ts
@@ -1,6 +1,5 @@
 import { goto } from "$app/navigation";
-import {
-  getRuntimeServiceListFilesQueryKey,
+import type {
   V1DeleteFileAndReconcileResponse,
   V1RenameFileAndReconcileResponse,
 } from "@rilldata/web-common/runtime-client";
@@ -9,6 +8,7 @@ import type { EntityType } from "@rilldata/web-local/common/data-modeler-state-s
 import { getNextEntityName } from "@rilldata/web-local/common/utils/getNextEntityId";
 import { dataModelerService } from "@rilldata/web-local/lib/application-state-stores/application-store";
 import { fileArtifactsStore } from "@rilldata/web-local/lib/application-state-stores/file-artifacts-store";
+import { invalidateAfterReconcile } from "@rilldata/web-local/lib/svelte-query/invalidation";
 import {
   getFileFromName,
   getLabel,
@@ -40,9 +40,7 @@ export async function renameFileArtifact(
   notifications.send({
     message: `renamed ${getLabel(type)} ${fromName} to ${toName}`,
   });
-  await queryClient.invalidateQueries(
-    getRuntimeServiceListFilesQueryKey(instanceId)
-  );
+  return invalidateAfterReconcile(queryClient, instanceId, resp);
 }
 
 export async function deleteFileArtifact(
@@ -67,10 +65,7 @@ export async function deleteFileArtifact(
     // Temporary until nodejs is removed
     await dataModelerService.dispatch("deleteEntity", [type, name]);
 
-    // TODO: update all entities based on affected path
-    return queryClient.invalidateQueries(
-      getRuntimeServiceListFilesQueryKey(instanceId)
-    );
+    return invalidateAfterReconcile(queryClient, instanceId, resp);
   } catch (err) {
     console.error(err);
   }
