+++ web-local/src/lib/svelte-query/dashboards.ts
@@ -1,4 +1,9 @@
-import { useRuntimeServiceListFiles } from "@rilldata/web-common/runtime-client";
+import {
+  useRuntimeServiceGetCatalogEntry,
+  useRuntimeServiceListFiles,
+  V1MetricsView,
+} from "@rilldata/web-common/runtime-client";
+import type { MetricsViewRequestFilter } from "@rilldata/web-local/common/rill-developer-service/MetricsViewActions";
 
 export function useDashboardNames(repoId: string) {
   return useRuntimeServiceListFiles(
@@ -19,3 +24,60 @@ export function useDashboardNames(repoId: string) {
     }
   );
 }
+
+export const useMetaQuery = <T = V1MetricsView>(
+  instanceId: string,
+  metricViewName: string,
+  selector?: (meta: V1MetricsView) => T
+) => {
+  return useRuntimeServiceGetCatalogEntry(instanceId, metricViewName, {
+    query: {
+      select: (data) =>
+        selector
+          ? selector(data?.entry?.metricsView)
+          : data?.entry?.metricsView,
+    },
+  });
+};
+
+export const useMetaMeasure = (
+  instanceId: string,
+  metricViewName: string,
+  measureName: string
+) =>
+  useMetaQuery(instanceId, metricViewName, (meta) =>
+    meta.measures?.find((measure) => measure.name === measureName)
+  );
+
+export const useMetaDimension = (
+  instanceId: string,
+  metricViewName: string,
+  dimensionName: string
+) =>
+  useMetaQuery(instanceId, metricViewName, (meta) =>
+    meta.dimensions?.find((dimension) => dimension.name === dimensionName)
+  );
+
+/**
+ * Returns a copy of the filter without the passed in dimension filters.
+ */
+export const getFilterForDimension = (
+  filters: MetricsViewRequestFilter,
+  dimensionName?: string
+) => {
+  if (!filters) return undefined;
+  return {
+    include: filters.include
+      .filter((dimensionValues) => dimensionName !== dimensionValues.name)
+      .map((dimensionValues) => ({
+        name: dimensionValues.name,
+        in: dimensionValues.in,
+      })),
+    exclude: filters.exclude
+      .filter((dimensionValues) => dimensionName !== dimensionValues.name)
+      .map((dimensionValues) => ({
+        name: dimensionValues.name,
+        in: dimensionValues.in,
+      })),
+  };
+};
