--- web-local/src/lib/svelte-query/queries/fetch-url.ts
@@ -1,18 +1,0 @@
-export async function fetchUrl(
-  base: string,
-  path: string,
-  method: string,
-  body?
-) {
-  const resp = await fetch(`${base}/api/v1/${path}`, {
-    method,
-    headers: { "Content-Type": "application/json" },
-    ...(body ? { body: JSON.stringify(body) } : {}),
-  });
-  const json = await resp.json();
-  if (!resp.ok) {
-    const err = new Error(json.messages[0].message);
-    return Promise.reject(err);
-  }
-  return json;
-}
